Fix /f say
Previously, the /f say conditional check disabled f say, probably because somebody planned to add a config line for it. Seems like that would be redundant so now if faction chat is enabled in config, so is /f say. Also moved /f say to "CHAT" section of command file instead of being in the "MEMBERS/OFFICERS/LEADER AND THEIR STATUSES" section