OMG, how did I forget about this?
BTW, Fixed /f say again.
The reason why /f say was still broken after I announced it was fixed, is because Silly old me, I put the code into the Chat section, forgetting to delete the /f say section when it wasn't suppose to be there. Basically, a human error. I put 2 of the same code in different sections, which then the plugin would get confused as to what code /f say should be in, so instead by default, it allowed the /f say code to be in the default section (Originally). So the fix to the code wasn't valid due to the 2 of the same code in that specific class. Also, allowing The prefs "AllowChat" to be true. If true, then /f say will also be enabled. Basically I should of deleted the /f say code (The unfixed one), and then should've announced its fix. Again, sorry for those who were hyped. But it should now be fixed. If not, please open a new issue.