Fixed some commands returning as true
For instance, there was no point in having "return true" when the command its self had already been set to either true or false. This would completely confuse the code into thinking return true, and return false. So I removed the ones which already had the true bits on the code, when that was needed, but the return true code wasn't needed, as it would be no need in having both codes in one, as it was pointless. So we just removed it. :P
BTW, I also customised /f say, as we fully fixed it, tested by BadderSubset. (A staff member on my server.)
Big shout out to him.
I believe /f overclaim should be fixed, as I believed this affected some parts of the command, which could potentially fix issue #1, but can't say for sure. Anyways, get testing. :P