Revamped faction messages
To make it more cleaner.
Also, made /f setwarp as a TO do list, as it currently isn't complete yet. It is still currently buggy, so for now, it will be disabled untill we fix it.
Cleaned up some code
Fixed /f plotinfo again. I could've sworn I edited that earlier.
Added more to do with $args[1] function, as everyone loves those functions in coding. ;3
Cleaned up some messages, and added some messages which made sense.
A ton of bug fixes. Go and test it yourself.
Bumped Plugin to v2.0.0.