Fixed /f leave
Not sure why this was added in the first place, but we've added a way so if you're not in a faction when using /f leave, it will now come up with a message saying "You are not in this faction" instead of leaving it out of the blue, even if you aren't in the faction. This bug was probably there for a long time, but no one never noticed it. Along with fixing /f leave crashes as well. Not sure if that was patched in this build however, it was patched in other builds of this plugin.