Fixed Leader issues (#29)
This fixes You, as the leader, to be able to bypass the /f promote and /f demote check, as iyou'd be able to bypass yourself to demote or promote yourself, although, you're a leader.
Also, big thanks to @SuperAlecs for reporting this issue first. Please test this out, and see if any issues or problems still assist or not, thanks.