Fixed #30
I haven't seen this bug, but a user has reported this, and it would've made sense if this was the fix, as the isInFaction part means that they're in a faction. Which clearly says You are not in a faction. And the return is true, which could've been false in that matter however, the ($sender->getName()) == false) { clarifies the function or code is false. But this should fix it. Please test it, and let me know if this fixes it or not.