Fixed /f promote and /f demote
We replaced $promotee and $demotee with $args[1], based on the user typing in that username, even if it's wrong. It'll simply copy, as to what you typed. We replaced them because $promotee and $demotee are broken. When you execute the command, instead of it coming up with the proper username, it'll come up as Player(1), which this build should fix that.